[REFACTOR] #13 - BanChanRepository 위치 domain -> data로 이동
