+++ app/src/main/java/com/example/banchan/di/RepositoryModule.kt
@@ -2,7 +2,7 @@ package com.example.banchan.di
 
 import com.example.banchan.data.repository.BanChanRepositoryImpl
 import com.example.banchan.data.source.remote.BanChanRemoteDataSource
-import com.example.banchan.domain.repository.BanChanRepository
+import com.example.banchan.data.repository.BanChanRepository
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -19,7 +19,7 @@ class RepositoryModule {
     fun provideBanChanRepository(
         banChanDataSource: BanChanRemoteDataSource,
         ioDispatcher: CoroutineDispatcher
-    ): BanChanRepository{
+    ): BanChanRepository {
         return BanChanRepositoryImpl(banChanDataSource, ioDispatcher)
     }
 }
