+++ app/src/main/java/com/example/banchan/domain/usecase/GetBestDishesUseCase.kt
@@ -1,7 +1,7 @@
 package com.example.banchan.domain.usecase
 
 import com.example.banchan.domain.model.BestModel
-import com.example.banchan.domain.repository.BanChanRepository
+import com.example.banchan.data.repository.BanChanRepository
 import javax.inject.Inject
 
 class GetBestDishesUseCase @Inject constructor(
